+++ moai_app/lib/main.dart
@@ -1,7 +1,11 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:moai_app/playground/playground_launcher.dart';
+import 'package:moai_app/services/huddle01/huddle_playground.dart';
+import 'package:moai_app/services/huddle01/template.dart';
 import 'package:moai_app/services/xmtp/xmtpinterface.dart';
 
 final navigatorKey = GlobalKey<NavigatorState>();
@@ -9,7 +13,9 @@ FlutterSecureStorage get storage => const FlutterSecureStorage();
 final gpc = ProviderContainer();
 
 void main() {
+  WidgetsFlutterBinding.ensureInitialized();
   MoaiXMTPInterface.instance.initalizeWalletProvider();
+  HttpOverrides.global = MyHttpOverrides();
   runApp(
     UncontrolledProviderScope(
       container: gpc,
@@ -33,7 +39,7 @@ class MoaiApplication extends StatelessWidget {
           useMaterial3: true,
         ),
         navigatorKey: navigatorKey,
-        home: const PlaygroundLauncher(),
+        home: const HuddlePlayground(),
       ),
     );
   }
@@ -49,5 +55,14 @@ class WelcomeScreen extends StatelessWidget {
         title: const Text('Welcome to Moai'),
       ),
     );
+  }
+}
+
+class MyHttpOverrides extends HttpOverrides {
+  @override
+  HttpClient createHttpClient(SecurityContext? context) {
+    return super.createHttpClient(context)
+      ..badCertificateCallback =
+          (X509Certificate cert, String host, int port) => true;
   }
 }