+++ moai_app/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 9.12.2023 [5:41]
+- [feat]: Huddle Engine Created
+- [feat]: Huddle01 Playground Created
+
 # 9.12.2023 [3:00]
 - [feat]: CustomWallet Integration done (Riverpod impl)
 - [feat]: XMTP Integration done (RiverPod impl)